Added courses page
Having difficulties with the footer not touching the bottom of the page.

The top image is not removed because when it is changed into a solid shape, the image on the index page also changes.

The bottom of the index page shows another image that isn't on the page.

Removed unnecessary tags used for understanding in the CSS file.